Rename Hanami::Router#namespace into #prefix
